Add Mod Manager button. Add Mod Manager main screen fragment. Move "Install Modrinth modpack" button to Mod Manager fragment. Change "Install .mrpack modpack" to "Install Modrinth modpack (mrpack)". Add "Download Mods" button (doesn't do anything right now). Changed default GUIScale in Minecraft to be 0 (auto) Changed Snooper's data collection thingy in Minecraft to be false
