+++ app_pojavlauncher/src/main/assets/components/lwjgl3/version
@@ -1,1 +1,1 @@
-1679697617912
+1679707343124
