+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/fragments/MainMenuFragment.java
@@ -15,6 +15,7 @@ import androidx.fragment.app.Fragment;
 
 
 import net.kdt.pojavlaunch.CustomControlsActivity;
+import net.kdt.pojavlaunch.ModManagerMain;
 import net.kdt.pojavlaunch.R;
 import net.kdt.pojavlaunch.Tools;
 import net.kdt.pojavlaunch.extra.ExtraConstants;
@@ -36,7 +37,7 @@ public class MainMenuFragment extends Fragment {
         Button mCustomControlButton = view.findViewById(R.id.custom_control_button);
         Button mInstallJarButton = view.findViewById(R.id.install_jar_button);
         Button mShareLogsButton = view.findViewById(R.id.share_logs_button);
-        Button mInstallMrpackButton = view.findViewById(R.id.install_mrpack_button);
+        Button mModManagerButton = view.findViewById(R.id.mod_manager_button);
 
         ImageButton mEditProfileButton = view.findViewById(R.id.edit_profile_button);
         Button mPlayButton = view.findViewById(R.id.play_button);
@@ -49,7 +50,8 @@ public class MainMenuFragment extends Fragment {
             return true;
         });
 
-        mInstallMrpackButton.setOnClickListener(v -> runMrpackInstaller());
+        // mModManagerButton.setOnClickListener(v -> {throw new RuntimeException("Bruh");});
+        mModManagerButton.setOnClickListener(v -> Tools.swapFragment(requireActivity(), ModManagerMain.class, ModManagerMain.TAG, true, null));
 
         mEditProfileButton.setOnClickListener(v -> Tools.swapFragment(requireActivity(), ProfileEditorFragment.class, ProfileEditorFragment.TAG, true, null));
 
@@ -70,12 +72,7 @@ public class MainMenuFragment extends Fragment {
 
     }
 
-    private void runMrpackInstaller() {
-        if (ProgressKeeper.getTaskCount() == 0)
-            Tools.installModpack(requireActivity());
-        else
-            Toast.makeText(requireContext(), R.string.tasks_ongoing, Toast.LENGTH_LONG).show();
-    }
+
 
     private void runInstallerWithConfirmation(boolean isCustomArgs) {
         if (ProgressKeeper.getTaskCount() == 0)
