+++ app_pojavlauncher/src/main/res/layout-land/fragment_launcher.xml
@@ -57,12 +57,12 @@
 
 
     <com.kdt.mcgui.LauncherMenuButton
-        android:id="@+id/install_mrpack_button"
+        android:id="@+id/mod_manager_button"
         android:layout_width="match_parent"
         android:layout_height="@dimen/_66sdp"
         android:background="?android:attr/selectableItemBackground"
         android:drawableStart="@drawable/modrinth_logo"
-        android:text="@string/main_install_mrpack"
+        android:text="@string/main_mod_manager"
 
         app:layout_constraintTop_toBottomOf="@id/share_logs_button" />
