+++ app_pojavlauncher/src/main/res/layout/fragment_mod_manager_main.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ModManagerMain">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <com.kdt.mcgui.LauncherMenuButton
+            android:id="@+id/install_mrpack_button"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/_66sdp"
+            android:background="?android:attr/selectableItemBackground"
+            android:drawableStart="@drawable/modrinth_logo"
+            android:text="@string/modmanager_install_mrpack" />
+
+
+        <com.kdt.mcgui.LauncherMenuButton
+            android:id="@+id/download_mods"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/_66sdp"
+            android:background="?android:attr/selectableItemBackground"
+            android:drawableStart="@drawable/download"
+            android:text="@string/modmanager_download_mods"
+            app:layout_constraintTop_toBottomOf="@id/install_mrpack_button" />
+
+        <View
+            android:id="@+id/_background_display_view"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:background="@color/background_bottom_bar"
+            android:translationZ="-1dp"
+
+            app:layout_constraintBottom_toBottomOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</FrameLayout>
