+++ app_pojavlauncher/src/main/res/values/strings.xml
@@ -359,7 +359,11 @@
     <!-- Community Edition (Strings)-->
     <string name="enable_forgesplash_title">Enable Forge splash screen</string>
     <string name="enable_forgesplash_description">Enables the Forge splash screen. WARNING: Can cause issues.</string>
-    <string name="main_install_curseforgemodpack">Install Curseforge modpack</string>
-    <string name="main_install_mrpack">Install .mrpack modpack</string>
     <string name="mcl_setting_renderer_gl4es115">GL4ES 1.1.5 - (all versions, fast)</string>
+
+    <!-- Mod Manager -->
+    <string name="main_mod_manager">Mod Manager</string>
+    <string name="modmanager_install_curseforgemodpack">Install Curseforge modpack</string>
+    <string name="modmanager_install_mrpack">Install Modrinth modpack (mrpack)</string>
+    <string name="modmanager_download_mods">Download mods</string>
 </resources>
